requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0

# Supabase
supabase==2.3.4

# Scraping
playwright==1.41.0

# Audio transcription
openai==1.35.0
deepgram-sdk==3.1.0

# HTTP client
httpx==0.25.2
aiofiles==23.2.1

# Data handling
pydantic==2.5.3
numpy==2.3.1
pydub==0.25.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0